Add Classes: 1. BasePages 2. CartPage 3. LogInPage 4. ProductPages 5. ProductsPage
Simple one test to confirm that all methods are working.
Added README.md
