+++ README.md
@@ -1,0 +1,7 @@
+## Checklist for Saucedemo [![SauceDemo](https://res.cloudinary.com/duauoz75o/image/upload/c_scale,w_135/v1620137689/Login_Bot_graphic.20658452_vymtmk.png)](https://www.saucedemo.com/)
+
+<img src="https://res.cloudinary.com/duauoz75o/image/upload/c_scale,w_135/v1620137689/Login_Bot_graphic.20658452_vymtmk.png" align="right" />
+| First Header  | Second Header |
+| ------------- | ------------- |
+| Content Cell  | Content Cell  |
+| Content Cell  | Content Cell  |
