+++ pom.xml
@@ -25,6 +25,12 @@
             <version>7.4.0</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>io.github.bonigarcia</groupId>
+            <artifactId>webdrivermanager</artifactId>
+            <version>4.4.1</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
     
 </project>
