+++ src/test/java/pages/LogInPage.java
@@ -1,0 +1,29 @@
+package pages;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+
+public class LogInPage extends BasePages {
+    public static final By USERNAME_INPUT = By.id("user-name");
+    public static final By PASSWORD_INPUT = By.id("password");
+    public static final By LOGIN_BUTTON = By.id("login-button");
+    public static final By ERROR_MASSAGE = By.cssSelector("[data-test=error]");
+
+    public LogInPage(WebDriver driver) {
+        super(driver);
+    }
+
+    public void open() {
+        driver.get("https://www.saucedemo.com/");
+    }
+
+    public void logIn(String user, String password) {
+        driver.findElement(USERNAME_INPUT).sendKeys(user);
+        driver.findElement(PASSWORD_INPUT).sendKeys(password);
+        driver.findElement(LOGIN_BUTTON).click();
+    }
+
+    public String getError() {
+        return driver.findElement(ERROR_MASSAGE).getText();
+    }
+}
